Create Home and Settings pages
