+++ src/components/Pomodoro/index.tsx
@@ -7,8 +7,6 @@ import { pomodoroDurations } from "../../library/pomodoroDurations";
 import { PomodoroStatus } from "./PomodoroStatus/index";
 import { PomodoroScreen } from "./PomodoroScreen/index";
 import { PomodoroControls } from "./PomodoroControls/index";
-import { SettingsButton } from "./styles";
-import { ReactComponent as IconSettings } from "./../../assets/icons/settings.svg";
 
 const Pomodoro = () => {
 	const [ pomodorosCompleted, setPomodorosCompleted ] = useState(0);
@@ -58,9 +56,6 @@ const Pomodoro = () => {
 
 	return (
 		<div>
-			<SettingsButton to="/settings">
-				<IconSettings />
-			</SettingsButton>
 			<PomodoroStatus
 				currentStatus={ currentStatus.title }
 				pomodorosCompleted={ pomodorosCompleted }
