+++ src/pages/Settings.tsx
@@ -1,0 +1,15 @@
+import { BackButton } from "./styles";
+import { ReactComponent as IconBackArrow } from "../assets/icons/back-arrow.svg";
+
+const Settings = () => {
+	return (
+		<>
+			<BackButton to="/">
+				<IconBackArrow />
+			</BackButton>
+			<h1> Settings Page! </h1>
+		</>
+	)
+}
+
+export { Settings };
