+++ src/pages/styles.ts
@@ -1,6 +1,17 @@
 import styled from "styled-components";
 import { Link } from "react-router-dom";
 
+const BackButton = styled(Link)`
+    position: absolute;
+    top: 20px;
+    left: 1rem;
+
+    svg {
+        width: 2.5rem;
+        height: 2.5rem;
+    }
+`;
+
 const SettingsButton = styled(Link)`
     position: absolute;
     top: 20px;
@@ -12,4 +23,4 @@ const SettingsButton = styled(Link)`
     }
 `;
 
-export { SettingsButton };
+export { BackButton, SettingsButton };
